(#16) changed fixture to only delete files it detects after tests completes. Avoids errors which prevents file deletions when only running test subsets

+++ tests/conftest.py
@@ -30,4 +30,5 @@ def test_cleanup():
     yield
     files = ["quvax.qu", "quvax.log", "sequences.txt", "test_felandscape_out.txt", "test_fegeneration_out.txt"]
     for file in files:
-        os.remove(file)
+        if os.path.exists(file):
+            os.remove(file)
